+++ hack/multi-node/user-data.sample
@@ -9,7 +9,7 @@ coreos:
         [Service]
         EnvironmentFile=/etc/environment
         Environment=KUBELET_IMAGE_URL=docker://k8s.gcr.io/hyperkube
-        Environment=KUBELET_IMAGE_TAG=v1.12.1
+        Environment=KUBELET_IMAGE_TAG=v1.13.3
         Environment="RKT_RUN_ARGS=--uuid-file-save=/var/cache/kubelet-pod.uuid \
           --volume var-lib-cni,kind=host,source=/var/lib/cni \
           --volume var-lib-kubelet,kind=host,source=/var/lib/kubelet \
