+++ hack/tests/conformance-test.sh
@@ -2,7 +2,7 @@
 set -euo pipefail
 
 CONFORMANCE_REPO=${CONFORMANCE_REPO:-github.com/kubernetes/kubernetes}
-CONFORMANCE_VERSION=${CONFORMANCE_VERSION:-v1.12.1}
+CONFORMANCE_VERSION=${CONFORMANCE_VERSION:-v1.13.3}
 
 usage() {
     echo "USAGE:"
