+++ pkg/asset/images.go
@@ -8,6 +8,6 @@ var DefaultImages = ImageVersions{
 	Calico:          "quay.io/calico/node:v3.0.3",
 	CalicoCNI:       "quay.io/calico/cni:v2.0.0",
 	CoreDNS:         "k8s.gcr.io/coredns:1.2.4",
-	Hyperkube:       "k8s.gcr.io/hyperkube:v1.12.1",
+	Hyperkube:       "k8s.gcr.io/hyperkube:v1.13.3",
 	PodCheckpointer: "quay.io/coreos/pod-checkpointer:83e25e5968391b9eb342042c435d1b3eeddb2be1",
 }
