+++ pkg/checkpoint/internal/README.md
@@ -1,6 +1,6 @@
 Copied from pkg/kubelet/util to prevent importing kubernetes
 
 ```
-wget https://raw.githubusercontent.com/kubernetes/kubernetes/v1.12.3/pkg/kubelet/util/util_unix.go
-wget https://raw.githubusercontent.com/kubernetes/kubernetes/v1.12.3/pkg/kubelet/util/util.go
+wget https://raw.githubusercontent.com/kubernetes/kubernetes/v1.13.3/pkg/kubelet/util/util_unix.go
+wget https://raw.githubusercontent.com/kubernetes/kubernetes/v1.13.3/pkg/kubelet/util/util.go
 ```
